API / configuration: added `actor_model_names`
This `Array` of `String`s allows to list the models that should be
handled in a special way when it comes to associations naming.
If any other model gets associated with one of these, the association
names won't include the name of the latter one.

It's much easier to learn that from example. We'll get `authors` instead
of `author_users` if `'User'` is in `actor_model_names` (which is a
default).